requirements.txt
Pillow
pdf2image
fastapi
uvicorn[standard]